added depedencies for express server CD
