+++ .gitignore
@@ -4,7 +4,8 @@
 */node_modules
 */.pnp
 *.pnp.js
-
+*/functions
+node_modules
 # testing
 */coverage
