+++ netlify.toml
@@ -1,0 +1,5 @@
+[build]
+  functions = "functions"
+
+[dev]
+  publish = "app"
