+++ package.json
@@ -1,0 +1,9 @@
+{
+  "devDependencies": {
+    "netlify-cli": "^12.10.0"
+  },
+  "scripts": {
+    "start": "./server/node_modules/.bin/netlify-lambda serve server",
+    "build": "./server/node_modules/.bin/netlify-lambda build server"
+  }
+}
